@@ -1,12 +1,13 @@
 name: photo_manager
 description: You can scan photos and albums. Only api, not have ui. you can use the api to create your image picker. or use photo
-version: 0.6.0-dev.7
+version: 1.1.4
 # author: caijinglong<cjl_spy@163.com>
 homepage: https://github.com/CaiJingLong/flutter_photo_manager
 
 environment:
   sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.10.0 <2.0.0"
+  flutter: ">=2.0.0"
+  
 dependencies:
   flutter:
     sdk: flutter
