@@ -12,7 +12,7 @@ class _DevelopingExampleState extends State<DevelopingExample> {
     return Scaffold(
       appBar: AppBar(),
       body: Container(
-        child: RaisedButton(
+        child: ElevatedButton(
           child: Text("Test title speed"),
           onPressed: () async {
             final start = DateTime.now();
@@ -23,10 +23,9 @@ class _DevelopingExampleState extends State<DevelopingExample> {
               List<AssetPathEntity> list = await PhotoManager.getAssetPathList(
                 type: RequestType.image,
               );
-              if (list != null)
-                for (AssetPathEntity path in list)
-                  imageList.addAll(await path.getAssetListRange(
-                      start: 0, end: path.assetCount));
+              for (AssetPathEntity path in list)
+                imageList.addAll(await path.getAssetListRange(
+                    start: 0, end: path.assetCount));
 
               if (imageList.isNotEmpty) {
                 imageList.shuffle();
