@@ -4,14 +4,14 @@ import 'package:image_scanner_example/widget/gallery_item_widget.dart';
 import 'package:provider/provider.dart';
 
 class GalleryListPage extends StatefulWidget {
-  const GalleryListPage({Key key}) : super(key: key);
+  const GalleryListPage({Key? key}) : super(key: key);
 
   @override
   _GalleryListPageState createState() => _GalleryListPageState();
 }
 
 class _GalleryListPageState extends State<GalleryListPage> {
-  PhotoProvider get provider => Provider.of<PhotoProvider>(context);
+  PhotoProvider get provider => context.watch<PhotoProvider>();
 
   @override
   Widget build(BuildContext context) {
