@@ -15,7 +15,8 @@ class NewHomePage extends StatefulWidget {
 }
 
 class _NewHomePageState extends State<NewHomePage> {
-  PhotoProvider get provider => Provider.of<PhotoProvider>(context);
+  PhotoProvider get readProvider => context.read<PhotoProvider>();
+  PhotoProvider get watchProvider => context.watch<PhotoProvider>();
 
   @override
   void initState() {
@@ -27,7 +28,7 @@ class _NewHomePageState extends State<NewHomePage> {
   @override
   Widget build(BuildContext context) {
     return ChangeNotifierBuilder(
-      value: provider,
+      value: watchProvider,
       builder: (_, __) => Scaffold(
         appBar: AppBar(
           title: Text("photo manager example"),
@@ -44,13 +45,14 @@ class _NewHomePageState extends State<NewHomePage> {
                 ),
               ],
             ),
-            _buildTypeChecks(provider),
+            _buildTypeChecks(watchProvider),
             _buildHasAllCheck(),
             _buildOnlyAllCheck(),
             _buildContainsEmptyCheck(),
+            _buildPathContainsModifiedDateCheck(),
             _buildPngCheck(),
             _buildNotifyCheck(),
-            _buildFilterOption(provider),
+            _buildFilterOption(watchProvider),
           ],
         ),
       ),
@@ -75,8 +77,8 @@ class _NewHomePageState extends State<NewHomePage> {
         child: CheckboxListTile(
           title: Text(typeText),
           value: currentType.containsType(type),
-          onChanged: (bool value) {
-            if (value) {
+          onChanged: (bool? value) {
+            if (value == true) {
               provider.changeType(currentType + type);
             } else {
               provider.changeType(currentType - type);
@@ -99,7 +101,7 @@ class _NewHomePageState extends State<NewHomePage> {
   }
 
   _scanGalleryList() async {
-    await provider.refreshGalleryList();
+    await readProvider.refreshGalleryList();
 
     final page = GalleryListPage();
 
@@ -110,9 +112,9 @@ class _NewHomePageState extends State<NewHomePage> {
 
   Widget _buildHasAllCheck() {
     return CheckboxListTile(
-      value: provider.hasAll,
+      value: watchProvider.hasAll,
       onChanged: (value) {
-        provider.changeHasAll(value);
+        readProvider.changeHasAll(value);
       },
       title: Text("hasAll"),
     );
@@ -120,9 +122,9 @@ class _NewHomePageState extends State<NewHomePage> {
 
   Widget _buildPngCheck() {
     return CheckboxListTile(
-      value: provider.thumbFormat == ThumbFormat.png,
+      value: watchProvider.thumbFormat == ThumbFormat.png,
       onChanged: (value) {
-        provider.changeThumbFormat();
+        readProvider.changeThumbFormat();
       },
       title: Text("thumb png"),
     );
@@ -130,9 +132,9 @@ class _NewHomePageState extends State<NewHomePage> {
 
   Widget _buildOnlyAllCheck() {
     return CheckboxListTile(
-      value: provider.onlyAll,
+      value: watchProvider.onlyAll,
       onChanged: (value) {
-        provider.changeOnlyAll(value);
+        readProvider.changeOnlyAll(value);
       },
       title: Text("onlyAll"),
     );
@@ -143,22 +145,31 @@ class _NewHomePageState extends State<NewHomePage> {
       return Container();
     }
     return CheckboxListTile(
-      value: provider.containsEmptyAlbum,
+      value: watchProvider.containsEmptyAlbum,
       onChanged: (value) {
-        provider.changeContainsEmptyAlbum(value);
+        readProvider.changeContainsEmptyAlbum(value);
       },
       title: Text("contains empty album(only iOS)"),
     );
   }
 
+  Widget _buildPathContainsModifiedDateCheck() {
+    return CheckboxListTile(
+      value: watchProvider.containsPathModified,
+      onChanged: (value) {
+        readProvider.changeContainsPathModified(value);
+      },
+      title: Text("contains path modified date"),
+    );
+  }
+
   Widget _buildNotifyCheck() {
     return CheckboxListTile(
-        value: provider.notifying,
+        value: watchProvider.notifying,
         title: Text("onChanged"),
         onChanged: (value) {
-          provider.notifying = value;
-
-          if (value) {
+          readProvider.notifying = value;
+          if (value == true) {
             PhotoManager.startChangeNotify();
           } else {
             PhotoManager.stopChangeNotify();
@@ -169,7 +180,7 @@ class _NewHomePageState extends State<NewHomePage> {
   void onChange(call) {}
 
   Widget _buildFilterOption(PhotoProvider provider) {
-    return RaisedButton(
+    return ElevatedButton(
       child: Text("Change filter options."),
       onPressed: () {
         Navigator.push(
@@ -185,8 +196,8 @@ class _NewHomePageState extends State<NewHomePage> {
   }
 }
 
-Widget buildButton(String text, Function function) {
-  return RaisedButton(
+Widget buildButton(String text, VoidCallback function) {
+  return ElevatedButton(
     child: Text(text),
     onPressed: function,
   );
