@@ -335,7 +335,7 @@ class PhotoManagerPlugin(
           val ids = call.argument<List<String>>("ids")!!
           if (belowSdk(29)) {
             deleteManager.deleteInApi28(ids)
-            resultHandler.reply(true)
+            resultHandler.reply(ids)
           } else if (IDBUtils.isAndroidR) {
             val uris = ids.map {
               photoManager.getUri(it)
@@ -361,7 +361,8 @@ class PhotoManagerPlugin(
             val image = call.argument<ByteArray>("image")!!
             val title = call.argument<String>("title") ?: ""
             val desc = call.argument<String>("desc") ?: ""
-            val entity = photoManager.saveImage(image, title, desc)
+            val relativePath = call.argument<String>("relativePath") ?: ""
+            val entity = photoManager.saveImage(image, title, desc, relativePath)
             if (entity == null) {
               resultHandler.reply(null)
               return@runOnBackground
@@ -380,7 +381,8 @@ class PhotoManagerPlugin(
             val imagePath = call.argument<String>("path")!!
             val title = call.argument<String>("title") ?: ""
             val desc = call.argument<String>("desc") ?: ""
-            val entity = photoManager.saveImage(imagePath, title, desc)
+            val relativePath = call.argument<String>("relativePath") ?: ""
+            val entity = photoManager.saveImage(imagePath, title, desc, relativePath)
             if (entity == null) {
               resultHandler.reply(null)
               return@runOnBackground
@@ -399,7 +401,8 @@ class PhotoManagerPlugin(
             val videoPath = call.argument<String>("path")!!
             val title = call.argument<String>("title")!!
             val desc = call.argument<String>("desc") ?: ""
-            val entity = photoManager.saveVideo(videoPath, title, desc)
+            val relativePath = call.argument<String>("relativePath") ?: ""
+            val entity = photoManager.saveVideo(videoPath, title, desc, relativePath)
             if (entity == null) {
               resultHandler.reply(null)
               return@runOnBackground
