@@ -20,7 +20,8 @@ data class AssetEntity(
         val orientation: Int,
         var lat: Double? = null,
         var lng: Double? = null,
-        val androidQRelativePath: String? = null
+        val androidQRelativePath: String? = null,
+        val mimeType:String? = null
 ) {
 
   fun getUri(): Uri {
