@@ -1,4 +1,4 @@
-#import <Flutter/Flutter.h>
+#import "PMImport.h"
 
 @interface ImageScannerPlugin : NSObject <FlutterPlugin>
 @property(nonatomic, strong) NSObject <FlutterPluginRegistrar> *registrar;
