@@ -1,13 +1,6 @@
-import 'package:flutter/material.dart';
-
 import 'type.dart';
 
 class ThumbOption {
-  final int width;
-  final int height;
-  final ThumbFormat format;
-  final int quality;
-
   const ThumbOption({
     required this.width,
     required this.height,
@@ -21,7 +14,7 @@ class ThumbOption {
     ThumbFormat format = ThumbFormat.jpeg,
     int quality = 95,
     DeliveryMode deliveryMode = DeliveryMode.opportunistic,
-    ResizeMode resizeMode = ResizeMode.none,
+    ResizeMode resizeMode = ResizeMode.fast,
     ResizeContentMode resizeContentMode = ResizeContentMode.fit,
   }) {
     return _IosThumbOption(
@@ -35,6 +28,11 @@ class ThumbOption {
     );
   }
 
+  final int width;
+  final int height;
+  final ThumbFormat format;
+  final int quality;
+
   Map<String, Object> toMap() {
     return {
       'width': width,
@@ -45,24 +43,14 @@ class ThumbOption {
   }
 
   void checkAssert() {
-    assert(width != null && height != null,
-        "The width and height must not be null.");
     assert(width > 0 && height > 0, "The width and height must better 0.");
     assert(quality > 0 && quality <= 100, "The quality must between 0 and 100");
-    _checkNotNull('format', format);
-  }
-
-  void _checkNotNull(String key, value) {
-    assert(value != null, 'The $key must not be null.');
   }
 }
 
 class _IosThumbOption extends ThumbOption {
-  final DeliveryMode deliveryMode;
-  final ResizeMode resizeMode;
-  final ResizeContentMode resizeContentMode;
 
-  _IosThumbOption({
+  const _IosThumbOption({
     required int width,
     required int height,
     ThumbFormat format = ThumbFormat.jpeg,
@@ -77,13 +65,9 @@ class _IosThumbOption extends ThumbOption {
           quality: quality,
         );
 
-  @override
-  void checkAssert() {
-    super.checkAssert();
-    _checkNotNull('deliveryMode', deliveryMode);
-    _checkNotNull('resizeMode', resizeMode);
-    _checkNotNull('resizeContentMode', resizeContentMode);
-  }
+  final DeliveryMode deliveryMode;
+  final ResizeMode resizeMode;
+  final ResizeContentMode resizeContentMode;
 
   @override
   Map<String, Object> toMap() {
